LayoutTests/platform/glib/imported/w3c/web-platform-tests/trusted-types/should-trusted-type-policy-creation-be-blocked-by-csp-002-expected.txt

CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'policy name' because it violates the following Content Security Policy directive: "trusted-types policy name"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'policy*name' because it violates the following Content Security Policy directive: "trusted-types policy*name"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'policy$name' because it violates the following Content Security Policy directive: "trusted-types policy$name"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'policy?name' because it violates the following Content Security Policy directive: "trusted-types policy?name"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'policy!name' because it violates the following Content Security Policy directive: "trusted-types policy!name"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: The value for Content Security Policy directive 'trusted-types' contains an invalid character: 'polï¿½tica'. Non-whitespace characters outside ASCII 0x21-0x7E must be percent-encoded, as described in RFC 3986, section 2.1: http://tools.ietf.org/html/rfc3986#section-2.1.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named '_TTP3_' because it violates the following Content Security Policy directive: "trusted-types _TTP1_	_TTP2_"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named '_TTP4_' because it violates the following Content Security Policy directive: "trusted-types _TTP1_	_TTP2_"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named '_TTP5_' because it violates the following Content Security Policy directive: "trusted-types _TTP1_	_TTP2_"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named '_TTP6_' because it violates the following Content Security Policy directive: "trusted-types _TTP1_	_TTP2_"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named '_TTP_' because it violates the following Content Security Policy directive: "trusted-types _TTP_*"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.

PASS valid tt-policy-name name "1"
PASS valid tt-policy-name name "abcdefghijklmnopqrstuvwxyz"
PASS valid tt-policy-name name "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
PASS valid tt-policy-name name "0123456789"
PASS valid tt-policy-name name "policy-name"
PASS valid tt-policy-name name "policy=name"
PASS valid tt-policy-name name "policy_name"
PASS valid tt-policy-name name "policy/name"
PASS valid tt-policy-name name "policy@name"
PASS valid tt-policy-name name "policy.name"
PASS valid tt-policy-name name "policy%name"
PASS valid tt-policy-name name "policy#name"
PASS valid tt-policy-name name "6xY/2x=3Y..."
PASS invalid tt-policy-name name "policy name"
PASS invalid tt-policy-name name "policy*name"
PASS invalid tt-policy-name name "policy$name"
PASS invalid tt-policy-name name "policy?name"
PASS invalid tt-policy-name name "policy!name"
FAIL invalid tt-policy-name name "política" assert_true: createPolicy() should throw a TypeError. expected true got false
FAIL directive "trusted-type _TTP1_%09_TTP2_%0A_TTP3_%0C_TTP4_%0D_TTP5_%20_TTP6_" (required-ascii-whitespace) assert_equals: expected null but got object "TypeError: Failed to execute 'createPolicy': Policy with name '_TTP3_' disallowed."
PASS invalid directive "trusted-type _TTP" (no ascii whitespace)
